rename & moved the negative tests to a separate class
