+++ TestForParser/TestParser.cs
@@ -4,17 +4,8 @@ using System.Runtime.InteropServices;
 
 namespace TestForParser
 {
-    public class TestParser
+    public class ParserExpressionTests
     {
-        [Theory]
-        [InlineData("5*6+")]
-        public void JSON(string expression)
-        {
-            ExpressionTree tree = ExpressionTree.create_tree(expression);
-            string json = tree.json_tree();
-            Assert.Equal(1, 1);
-        }
-
         [Theory]
         [InlineData("x1+x2", new double[] { 1, 2 }, 3.0)]
         [InlineData("x1 - x2", new double[] { 5.0, 8.0 }, -3.0)]
@@ -99,81 +90,6 @@ namespace TestForParser
             Assert.Equal(result, actual);
         }
     }
-
-    public class TestParserExceptions
-    {
-        private void ExceptionEvaluteCatchingCheker(string expression, double[]? variables, string errorMessage)
-        {
-            var tree = ExpressionTree.create_tree(expression);
-
-            var exception = Assert.Throws<ApplicationException>(() => tree.evaluate([.. variables]));
-            Assert.Equal(errorMessage, exception.Message);
-        }
-
-        private void ExceptionCreateCatchingCheker(string expression, double[]? variables, string errorMessage)
-        {
-
-            var exception = Assert.Throws<ApplicationException>(() => ExpressionTree.create_tree(expression));
-            Assert.Equal(errorMessage, exception.Message);
-        }
-
-        [Theory]
-        [InlineData("x1/x2", new double[] { 10.0, 0.0 }, "division by zero.")]
-        [InlineData("x1/(x2-x3)", new double[] { 10.0, 5.0, 5.0 }, "division by zero.")]
-        [InlineData("1/0", new double[] {}, "division by zero.")]
-        public void DivideByZeroException(string expression, double[] variables, string errorMessage)
-            => ExceptionEvaluteCatchingCheker(expression, variables, errorMessage);
-
-        [Theory]
-        [InlineData("x1+6", new double[] { 50.0, 2.0 }, "The number of variables is incorrect")]
-        [InlineData("x1+x2", new double[] { 50.0, 2.0, 3.5 }, "The number of variables is incorrect")]
-        [InlineData("x1-x2", new double[] { 50.0}, "The number of variables is incorrect")]
-        [InlineData("x1+x2-x3", new double[] {}, "The number of variables is incorrect")]
-        public void IncorrectVariablesNumbers(string expression, double[] variables, string errorMessage)
-            => ExceptionEvaluteCatchingCheker(expression, variables, errorMessage);
-
-        [Theory]
-        [InlineData("x y", null, "Invalid expression string")]
-        [InlineData("2 3 +", null, "Invalid expression string")]
-        [InlineData("x + * y", null, "Invalid expression string")]
-        [InlineData("a / / b", null, "Invalid expression string")]
-        [InlineData("(x + y", null, "Invalid expression string")]
-        [InlineData("a * (b + c))", null, "Invalid expression string")]
-        [InlineData("()", null, "Invalid expression string")]
-        [InlineData("x + ()", null, "Invalid expression string")]
-        public void SyntaxException(string expression, double[]? variables, string errorMessage)
-            => ExceptionCreateCatchingCheker(expression, variables, errorMessage);
-
-        [Theory]
-        [InlineData(".", null, "Invalid expression string")]
-        [InlineData("123.", null, "Invalid expression string")]
-        [InlineData("1.2.3", null, "Invalid expression string")]
-        [InlineData("1#2", null, "Invalid expression string")]
-        [InlineData("a$ + b@", null, "Invalid expression string")]
-        [InlineData("", null, "Invalid expression string")] 
-        public void IncorrectNumbers(string expression, double[]? variables, string errorMessage)
-            => ExceptionCreateCatchingCheker(expression, variables, errorMessage);
-
-        [Theory]
-        [InlineData("foo(x)", null, "Invalid expression string")] 
-        [InlineData("bar(1, 2)", null, "Invalid expression string")]
-        [InlineData("sine(x)", null, "Invalid expression string")] 
-        [InlineData("sqrtt(y)", null, "Invalid expression string")]
-        public void IncorrectNamesAndFunction(string expression, double[]? variables, string errorMessage)
-            => ExceptionEvaluteCatchingCheker(expression, variables, errorMessage);
-
-        [Theory]
-        [InlineData("x1+x2", null, "Invalid expression string")]
-        public void ArgumentNullException(string expression, double[]? variables, string errorMessage)
-        {
-            var tree = ExpressionTree.create_tree(expression);
-
-            var exception = Assert.Throws<NullReferenceException>(() => tree.evaluate([.. variables]));
-            Assert.Equal(errorMessage, exception.Message);
-        }
-
-    }
-
 }
